# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/match/{id}": {
            "get": {
                "description": "Get match details by given match ID",
                "tags": [
                    "match"
                ],
                "summary": "Retrieve match details by match ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Match"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/matches": {
            "get": {
                "description": "Get matches filtered by their status (live, completed or scheduled) and sorted by round",
                "tags": [
                    "matches"
                ],
                "summary": "Retrieve matches by status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match Status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Match"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Match": {
            "type": "object",
            "properties": {
                "awayPlayerScore": {
                    "type": "integer"
                },
                "homePlayerScore": {
                    "type": "integer"
                },
                "matchID": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "round": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}